requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
Pillow>=8.3.0
tensorboard>=2.6.0
tqdm>=4.62.0
numpy>=1.21.0
matplotlib>=3.4.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
pydantic==2.4.2
starlette==0.27.0
scikit-learn>=0.24.0
seaborn>=0.11.0 